Refactor tests for SingleGeneratorTrainer: consolidate train_step and evaluate_step tests into a single file and enhance organization
